Remove jekyll_base/
Before, jekyll_base/ was copied to jekyll/ and data-generated files were generated into it. This made it tedious to develop the site because changes made to non-data files in jekyll/ were not tracked and had to be copied to jekyll_base/. Now, data-generated files and directories are listed in .gitignore.