Use _data for gene and locus
The gene and locus pages are now laid out in _layouts and load their table data from _data. The data for the trait associations table on the gene page are still on the gene markdown pages, though.

Gene links are also generated by the liquid template instead of included in the data. Gene IDs are shown when gene symbols are not available. Fixes #5

Project links now highlight the corresponding row in the project table.