Compute gene assoc tables in R and reformat
Builds model-trait association table directly in build_gene_pages.R instead of calling a shell script for each trait.

This allows reordering of models, which must correspond between the model table and association table. Now, they are ordered by decreasing mean Z^2.

Significant associations in the association table are now bolded. Fixes #6

Colors in the association table are now on a continuous gradient instead of a set number of shades.